Change package structure to DDD style
