+++ jutrack-cli/src/main/java/com/github/ibiber/jutrack/CommandLineParameterHandler.java
@@ -8,14 +8,15 @@ import java.time.LocalDate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import com.github.ibiber.jutrack.data.JiraQueryParmeter;
-import com.github.ibiber.jutrack.util.Strings;
-import com.github.ibiber.jutrack.util.Throwables;
+import com.github.ibiber.common.general.Strings;
+import com.github.ibiber.common.general.Throwables;
+import com.github.ibiber.jutrack.external.DefaultValueProviderService;
+import com.github.ibiber.jutrack.external.data.JiraQueryParmeter;
 
 @Component
 public class CommandLineParameterHandler {
 	@Autowired
-	private DefaultValueProvider defaultValue;
+	private DefaultValueProviderService defaultValue;
 
 	public JiraQueryParmeter getParameter(String... args) {
 		JiraQueryParmeter.Builder builder = new JiraQueryParmeter.Builder();
