+++ jutrack-core/src/main/java/com/github/ibiber/common/general/PropagatedException.java
@@ -1,0 +1,7 @@
+package com.github.ibiber.common.general;
+
+public class PropagatedException extends RuntimeException {
+	public PropagatedException(Throwable cause) {
+		super(cause);
+	}
+}
