+++ jutrack-core/src/main/java/com/github/ibiber/common/general/Strings.java
@@ -1,0 +1,10 @@
+package com.github.ibiber.common.general;
+
+public class Strings {
+	public static boolean isNullOrEmpty(String string) {
+		if (string == null) {
+			return true;
+		}
+		return string.isEmpty();
+	}
+}
