+++ jutrack-core/src/main/java/com/github/ibiber/common/general/Throwables.java
@@ -1,0 +1,7 @@
+package com.github.ibiber.common.general;
+
+public class Throwables {
+	public static PropagatedException propagate(Throwable e) {
+		throw new PropagatedException(e);
+	}
+}
