--- jutrack-core/src/main/java/com/github/ibiber/jutrack/GetIssueResultItemPresenter.java
@@ -1,10 +1,0 @@
-package com.github.ibiber.jutrack;
-
-import java.util.stream.Stream;
-
-import com.github.ibiber.jutrack.data.JiraQueryResultItem;
-import com.github.ibiber.jutrack.data.JiraQueryParmeter;
-
-public interface GetIssueResultItemPresenter {
-	public void presentResults(JiraQueryParmeter parameter, Stream<JiraQueryResultItem> resultStream);
-}
