+++ jutrack-core/src/main/java/com/github/ibiber/jutrack/application/ApplicationVersionProvider.java
@@ -1,0 +1,17 @@
+package com.github.ibiber.jutrack.application;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.PropertySource;
+import org.springframework.core.env.Environment;
+import org.springframework.stereotype.Component;
+
+@PropertySource("classpath:/juTrackBuild.properties")
+@Component
+public class ApplicationVersionProvider {
+	@Autowired
+	private Environment env;
+
+	public String getApplicationVersion() {
+		return env.getProperty("application.version");
+	}
+}
