--- jutrack-core/src/main/java/com/github/ibiber/jutrack/data/Credentials.java
@@ -1,18 +1,0 @@
-package com.github.ibiber.jutrack.data;
-
-public class Credentials {
-    public final String userName;
-    public final String password;
-
-    public Credentials(String userName, String password) {
-        super();
-        this.userName = userName;
-        this.password = password;
-    }
-
-    @Override
-    public String toString() {
-        return "Credentials [userName=" + userName + ", password set=" + (password != null && password.length() > 0)
-                + "]";
-    }
-}
