+++ jutrack-core/src/main/java/com/github/ibiber/jutrack/domain/data/jira/Author.java
@@ -1,0 +1,30 @@
+package com.github.ibiber.jutrack.domain.data.jira;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class Author {
+	public final String name;
+	public final String emailAddress;
+	public final String displayName;
+	public final String active;
+
+	/** Constructor needed by framework */
+	Author() {
+		this(null, null, null, null);
+	}
+
+	public Author(String name, String emailAddress, String displayName, String active) {
+		super();
+		this.name = name;
+		this.emailAddress = emailAddress;
+		this.displayName = displayName;
+		this.active = active;
+	}
+
+	@Override
+	public String toString() {
+		return "Author [name=" + name + ", emailAddress=" + emailAddress + ", displayName=" + displayName + ", active="
+		        + active + "]";
+	}
+}
