+++ jutrack-core/src/main/java/com/github/ibiber/jutrack/domain/data/jira/IssueFields.java
@@ -1,0 +1,14 @@
+package com.github.ibiber.jutrack.domain.data.jira;
+
+public class IssueFields {
+    public final String summary;
+
+    /** Constructor needed by framework */
+    IssueFields() {
+        this(null);
+    }
+
+    public IssueFields(String summary) {
+        this.summary = summary;
+    }
+}
