+++ jutrack-core/src/main/java/com/github/ibiber/jutrack/external/DefaultValueProviderService.java
@@ -1,0 +1,11 @@
+package com.github.ibiber.jutrack.external;
+
+public interface DefaultValueProviderService {
+	String getJiraRootUrlPropertyName();
+
+	String getJiraRootUrl();
+
+	String getUserName();
+
+	int dayRange();
+}
