+++ jutrack-core/src/main/java/com/github/ibiber/jutrack/external/JiraIssuesProcessorService.java
@@ -1,0 +1,9 @@
+package com.github.ibiber.jutrack.external;
+
+import com.github.ibiber.jutrack.external.data.JiraQueryParmeter;
+
+public interface JiraIssuesProcessorService {
+
+	void getIssues(JiraQueryParmeter parameter, GetIssueResultItemPresenter presenter);
+
+}
