--- jutrack-core/src/main/java/com/github/ibiber/jutrack/util/ApplicationVersionProvider.java
@@ -1,17 +1,0 @@
-package com.github.ibiber.jutrack.util;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.PropertySource;
-import org.springframework.core.env.Environment;
-import org.springframework.stereotype.Component;
-
-@PropertySource("classpath:/juTrackBuild.properties")
-@Component
-public class ApplicationVersionProvider {
-	@Autowired
-	private Environment env;
-
-	public String getApplicationVersion() {
-		return env.getProperty("application.version");
-	}
-}
