--- jutrack-core/src/main/java/com/github/ibiber/jutrack/util/PropagatedException.java
@@ -1,7 +1,0 @@
-package com.github.ibiber.jutrack.util;
-
-public class PropagatedException extends RuntimeException {
-	public PropagatedException(Throwable cause) {
-		super(cause);
-	}
-}
