--- jutrack-core/src/main/java/com/github/ibiber/jutrack/util/Strings.java
@@ -1,10 +1,0 @@
-package com.github.ibiber.jutrack.util;
-
-public class Strings {
-	public static boolean isNullOrEmpty(String string) {
-		if (string == null) {
-			return true;
-		}
-		return string.isEmpty();
-	}
-}
