--- jutrack-core/src/main/java/com/github/ibiber/jutrack/util/Throwables.java
@@ -1,7 +1,0 @@
-package com.github.ibiber.jutrack.util;
-
-public class Throwables {
-	public static PropagatedException propagate(Throwable e) {
-		throw new PropagatedException(e);
-	}
-}
