+++ jutrack-core/src/test/java/com/github/ibiber/common/http/UrlParameterBuilderTest.java
@@ -1,0 +1,54 @@
+package com.github.ibiber.common.http;
+
+import static org.hamcrest.Matchers.is;
+import static org.junit.Assert.assertThat;
+
+import org.junit.Before;
+import org.junit.Test;
+
+import com.github.ibiber.common.http.UrlParameterBuilder;
+
+public class UrlParameterBuilderTest {
+	private UrlParameterBuilder testee;
+
+	@Before
+	public void setUp() {
+		testee = new UrlParameterBuilder();
+	}
+
+	@Test
+	public void shouldBuildWithoutParameter() {
+		assertThat(testee.build(), is(""));
+	}
+
+	@Test
+	public void shouldBuildWithOneParameter() {
+		testee.add("k5", "v5");
+		assertThat(testee.build(), is("k5=v5"));
+	}
+
+	@Test
+	public void shouldBuildWithEmptyKey() {
+		testee.add("", "v5");
+		assertThat(testee.build(), is("=v5"));
+	}
+
+	@Test
+	public void shouldBuildWithEmptyValue() {
+		testee.add("k5", "");
+		assertThat(testee.build(), is("k5="));
+	}
+
+	@Test
+	public void shouldBuildWithManyParametersAsTheyWereAdded() {
+		testee.add("k5", "v5").add("k1", "v1");
+		assertThat(testee.build(), is("k5=v5&k1=v1"));
+	}
+
+	@Test
+	public void shouldBuildWithSameKeys() {
+		testee.add("k5", "v5").add("k1", "v1").add("k5", "v1");
+		assertThat(testee.build(), is("k5=v5&k1=v1&k5=v1"));
+	}
+
+}
