+++ jutrack-core/src/test/java/com/github/ibiber/jutrack/ModelIntegrationTest.java
@@ -30,9 +30,14 @@ import org.springframework.test.web.client.MockRestServiceServer;
 import org.springframework.test.web.client.match.MockRestRequestMatchers;
 import org.springframework.test.web.client.response.MockRestResponseCreators;
 
-import com.github.ibiber.commons.http.RestServiceQuery;
-import com.github.ibiber.jutrack.data.JiraQueryParmeter;
-import com.github.ibiber.jutrack.data.JiraQueryResultItem;
+import com.github.ibiber.common.http.RestServiceQuery;
+import com.github.ibiber.jutrack.domain.IssuesFilter;
+import com.github.ibiber.jutrack.domain.JiraIssuesProcessor;
+import com.github.ibiber.jutrack.domain.JiraIssuesQueryExecutor;
+import com.github.ibiber.jutrack.external.GetIssueResultItemPresenter;
+import com.github.ibiber.jutrack.external.JiraIssuesProcessorService;
+import com.github.ibiber.jutrack.external.data.JiraQueryParmeter;
+import com.github.ibiber.jutrack.external.data.JiraQueryResultItem;
 
 @RunWith(SpringRunner.class)
 @RestClientTest(JiraIssuesProcessor.class)
@@ -41,7 +46,7 @@ public class ModelIntegrationTest {
 	@Autowired
 	private TestGetIssueResultItemPresenter presenter;
 	@Autowired
-	private JiraIssuesProcessor testee;
+	private JiraIssuesProcessorService testee;
 	@Autowired
 	private MockRestServiceServer mockServer;
 
@@ -58,7 +63,7 @@ public class ModelIntegrationTest {
 
 		testee.getIssues(new JiraQueryParmeter("http://localhost:8080", "user_01", "anyPassword",
 		        LocalDate.from(DateTimeFormatter.ofPattern("yyyy-MM-dd").parse("2017-08-01")),
-		        LocalDate.from(DateTimeFormatter.ofPattern("yyyy-MM-dd").parse("2017-08-29"))));
+		        LocalDate.from(DateTimeFormatter.ofPattern("yyyy-MM-dd").parse("2017-08-29"))), presenter);
 
 		mockServer.verify();
 
@@ -89,9 +94,9 @@ public class ModelIntegrationTest {
 	@Configuration
 	public static class Config {
 		@Bean
-		public JiraIssuesProcessor jiraIssuesProcessor(JiraIssuesQueryExecutor jiraIssuesQueryExecutor,
-		        IssuesFilter<JiraQueryResultItem> filter, GetIssueResultItemPresenter presenter) {
-			return new JiraIssuesProcessor(jiraIssuesQueryExecutor, filter, presenter);
+		public JiraIssuesProcessorService jiraIssuesProcessor(JiraIssuesQueryExecutor jiraIssuesQueryExecutor,
+		        IssuesFilter<JiraQueryResultItem> filter) {
+			return new JiraIssuesProcessor(jiraIssuesQueryExecutor, filter);
 		}
 
 		@Bean
